make download app own ui component
